+++ src/tools/rerunWorkflow/handler.ts
@@ -1,14 +1,15 @@
 import { ToolCallback } from '@modelcontextprotocol/sdk/server/mcp.js';
-import { rerunWorkflowFromFailedInputSchema } from './inputSchema.js';
+import { rerunWorkflowInputSchema } from './inputSchema.js';
 import { getCircleCIClient } from '../../clients/client.js';
 
-export const rerunWorkflowFromFailed: ToolCallback<{
-  params: typeof rerunWorkflowFromFailedInputSchema;
+export const rerunWorkflow: ToolCallback<{
+  params: typeof rerunWorkflowInputSchema;
 }> = async (args) => {
   const { workflowId } = args.params;
   const circleci = getCircleCIClient();
-  const newWorkflow = await circleci.rerunWorkflow.rerunWorkflowFromFailed({
+  const newWorkflow = await circleci.rerunWorkflow.rerunWorkflow({
     workflowId,
+    fromFailed: true,
   });
   return {
     content: [